[CL-389] Exclude end slot label content from truncation
