+++ libs/components/src/form-control/form-control.module.ts
@@ -4,11 +4,11 @@ import { SharedModule } from "../shared";
 
 import { FormControlComponent } from "./form-control.component";
 import { BitHintComponent } from "./hint.component";
-import { BitLabel } from "./label.directive";
+import { BitLabel } from "./label.component";
 
 @NgModule({
-  imports: [SharedModule],
-  declarations: [FormControlComponent, BitLabel, BitHintComponent],
+  imports: [SharedModule, BitLabel],
+  declarations: [FormControlComponent, BitHintComponent],
   exports: [FormControlComponent, BitLabel, BitHintComponent],
 })
 export class FormControlModule {}
